Enable LowRes MV's flag when using NativeAA + OutputScaling

+++ OptiScaler/Util.cpp
@@ -597,6 +597,6 @@ void Util::GetDeviceRemovedReason(ID3D12Device* pDevice)
         break;
 
     default:
-        LOG_ERROR("Device removed reason: Unknown ({:X})", reason);
+        LOG_ERROR("Device removed reason: Unknown ({:X})", (UINT) reason);
     }
 }
